requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
python-multipart==0.0.6
pandas==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
openpyxl==3.1.2
fastapi-pagination==0.12.7
psycopg2-binary==2.9.9
